CI stability + warning removal

+++ Editor/CustomDrawers/Odin/WInLineEditorOdinDrawer.cs
@@ -111,7 +111,7 @@ namespace WallstopStudios.UnityHelpers.Editor.CustomDrawers
             }
 
             object parent = property.Parent?.ValueEntry?.WeakSmartValue;
-            int parentId = 0;
+            long parentId = 0;
             if (parent is Object unityObject)
             {
                 parentId = unityObject.GetUnityObjectId();
